Split the Monitors into separate files
Move the factory function into the NodeSpec; it appears to be impossible
to sensible resolve a circular import in JS.
